+++ src/App.js
@@ -4,6 +4,7 @@ import { RecoilRoot } from "recoil";
 import ScrollRestoration from "components/Skeleton/ScrollRestoration";
 import VirtualKeyboardDetector from "components/Skeleton/VirtualKeyboardDetector";
 import ChannelTalkProvider from "components/Skeleton/ChannelTalkProvider";
+import GoogleAnalyticsProvier from "components/Skeleton/GoogleAnalyticsProvier";
 import I18nextProvider from "lang/I18nextProvider";
 import AlertProvider from "components/Skeleton/AlertProvider";
 import CSSVariablesProvider from "components/Skeleton/CSSVariablesProvider";
@@ -31,6 +32,7 @@ const App = () => {
           <ScrollRestoration />
           <VirtualKeyboardDetector />
           <ChannelTalkProvider />
+          <GoogleAnalyticsProvier />
           <I18nextProvider />
           <AlertProvider />
           <CSSVariablesProvider />
